Move blacs_connector.h to module_external
