+++ source/source_base/parallel_2d.h
@@ -4,7 +4,7 @@
 #include <cstdint>
 #include <vector>
 
-#include "source_base/blacs_connector.h"
+#include "source_base/module_external/blacs_connector.h"
 
 /// @brief  This class packs the basic information of
 /// 2D-block-cyclic parallel distribution of an arbitrary matrix.
