+++ source/source_hsolver/diago_cusolver.cpp
@@ -1,6 +1,6 @@
 #include "diago_cusolver.h"
 
-#include "source_base/blacs_connector.h"
+#include "source_base/module_external/blacs_connector.h"
 #include "source_base/global_variable.h"
 #include "source_base/scalapack_connector.h"
 #include "source_base/timer.h"
