+++ source/source_hsolver/diago_elpa.cpp
@@ -2,7 +2,7 @@
 
 #include "source_io/module_parameter/parameter.h"
 #include "module_genelpa/elpa_solver.h"
-#include "source_base/blacs_connector.h"
+#include "source_base/module_external/blacs_connector.h"
 #include "source_base/global_variable.h"
 #include "source_base/lapack_connector.h"
 #include "source_base/scalapack_connector.h"
