+++ source/source_hsolver/diago_pxxxgvx.cpp
@@ -1,6 +1,6 @@
 #include "diago_pxxxgvx.h"
 
-#include "source_base/blacs_connector.h"
+#include "source_base/module_external/blacs_connector.h"
 #include "source_base/scalapack_connector.h"
 
 #include <complex>
