+++ source/source_hsolver/module_genelpa/utils.cpp
@@ -1,6 +1,6 @@
 #include "utils.h"
 
-#include "source_base/blacs_connector.h"
+#include "source_base/module_external/blacs_connector.h"
 #include "source_base/scalapack_connector.h"
 
 #include <complex>
