+++ source/source_lcao/module_rt/test/tddft_test.cpp
@@ -3,7 +3,7 @@
 #include <gtest/gtest.h>
 #include <mpi.h>
 
-#include "source_base/blacs_connector.h"
+#include "source_base/module_external/blacs_connector.h"
 #include "source_base/scalapack_connector.h"
 #include "source_basis/module_ao/parallel_orbitals.h"
